requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
attrs==22.2.0
boto3==1.26.61
botocore==1.29.61
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
ConfigArgParse==1.5.3
contourpy==1.0.7
cycler==0.11.0
exceptiongroup==1.1.0
Flask==2.2.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
fonttools==4.38.0
gevent==22.10.2
geventhttpclient==2.0.8
greenlet==2.0.1
idna==3.4
iniconfig==2.0.0
invokust==0.76
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.4
locust==2.2.1
MarkupSafe==2.1.2
matplotlib==3.6.3
msgpack==1.0.4
numpy==1.24.1
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pluggy==1.0.0
psutil==5.9.4
pyparsing==3.0.9
pytest==7.2.1
python-dateutil==2.8.2
pytz==2022.7.1
pyzmq==25.0.0
requests==2.28.2
roundrobin==0.0.4
s3transfer==0.6.0
seaborn==0.12.2
six==1.16.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
zope.event==4.6
zope.interface==5.5.2
